Add files via upload
This template was created as a way to practice the concepts and techniques studied in HTML and CSS. It features an organized structure, with separate files for HTML and CSS, which makes the customization process much easier and more intuitive.

In addition, the registration page template has an attractive and professional appearance, with a harmonious combination of colors, fonts and design elements. The UI is responsive which means the form can be accessed on different screen sizes.

I created this repository as a way to practice my studies in HTML and CSS, I improved my skills, from creating layouts to styling elements.